Some auth stuff is working. Moving on...
