+++ frontend/.env.example
@@ -1,0 +1,3 @@
+VITE_AUTH_DOMAIN=<domain>
+VITE_AUTH_CLIENT=<client id>
+VITE_AUTH_SECRET=<secret>
