--- frontend/src/Profile.tsx
@@ -1,27 +1,0 @@
-import { useAuth0, User } from "@auth0/auth0-react";
-import React from "react";
-
-const Profile = () => {
-  const { user, isAuthenticated, isLoading } = useAuth0();
-
-  if (isLoading) {
-    return <div>Loading ...</div>;
-  }
-
-  // Hmm
-  if (!user) {
-    return;
-  }
-
-  return (
-    isAuthenticated && (
-      <div>
-        <img src={user.picture} alt={user.name} />
-        <h2>{user.name}</h2>
-        <p>{user.email}</p>
-      </div>
-    )
-  );
-};
-
-export default Profile;
